article/serializers.py
from rest_framework import serializers

# Custom imports
from .models import Article


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = "__all__"