requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.34.122
botocore==1.34.122
Django==4.2.13
djangorestframework==3.15.1
ecdsa==0.19.0
jmespath==1.0.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
urllib3==1.26.18
